project is done

+++ ft_printf.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
+/*   By: dbredykh <dbredykh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/10 19:03:46 by dbredykh          #+#    #+#             */
-/*   Updated: 2023/05/12 15:43:05 by marvin           ###   ########.fr       */
+/*   Updated: 2023/05/14 12:37:28 by dbredykh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,9 +30,9 @@ size_t	ft_types(const char *str, size_t i, va_list param)
 	else if (str[i] == 'p')
 		count += ft_putvoid(va_arg(param, void *));
 	if (str[i] == 'x')
-		count += ft_puthexa(va_arg(param, size_t), 'x');
+		count += ft_puthexa(va_arg(param, unsigned int), 'x');
 	if (str[i] == 'X')
-		count += ft_puthexa(va_arg(param, size_t), 'X');
+		count += ft_puthexa(va_arg(param, unsigned int), 'X');
 	return (count);
 }
